refactor pool registration tests
